Update input_data.rst
Version 0.1.0
Description of the necessary input data for creating a scenario. 
Input data is divided between mandatory and optional data. 
Most of the data is described, however "decentralised heat" and "storages" are not well documented since in the first I am not sure if I got 100% the idea and in the second it is necessary to evaluate other types of storages.